fix: マイページが2つ表示される問題を修正
