+++ src/components/pages/MyPage/UserMenu/index.tsx
@@ -11,31 +11,39 @@ import {
 } from '@chakra-ui/react';
 import { HiUser, HiOutlineLogout, HiOutlineLogin } from 'react-icons/hi';
 import { supabase } from '@utils/supabaseClient';
-import { useContext } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
+import { UserDataContext } from '@pages/_app';
 import HeaderUserIcon from './HeaderUserIcon';
-import { type UserContextType, UserDataContext } from '@pages/_app';
 
 const UserMenu: React.FC = () => {
+  const [isGuest, setIsGuest] = useState<boolean>(false);
+
+  const contextData = useContext(UserDataContext);
+  const { isLoggedIn, userData } = contextData;
+
   const router = useRouter();
   const toast = useToast();
 
-  const { userData, authUser, isLoggedIn }: UserContextType =
-    useContext(UserDataContext);
+  useEffect(() => {
+    //ゲストログイン用アカウントをisGuestとして設定
+    if (userData?.id === process.env.NEXT_PUBLIC_GUEST_USER_ID) {
+      setIsGuest(true);
+    }
+  }, [userData]);
+
   const { avatar_url } = userData || {};
 
-  // ゲストログイン判定
-  const guestUserId = process.env.NEXT_PUBLIC_GUEST_USER_ID;
-  const isGuest = authUser?.id === guestUserId;
+  if (!userData) {
+    return null;
+  }
 
   const handleLogout = async () => {
     try {
       const { error } = await supabase.auth.signOut();
-
       if (error) {
         throw error;
       }
-
       toast({
         title: 'ログアウトが完了しました',
         status: 'success',
@@ -51,66 +59,49 @@ const UserMenu: React.FC = () => {
     }
   };
 
+  const handleClickUserMenu = () => {
+    if (!isLoggedIn) {
+      toast({
+        title: 'ログインされていない場合、マイページはご利用いただけません',
+        status: 'error',
+        duration: 6000,
+        position: 'top',
+        isClosable: true,
+      });
+    } else if (isGuest) {
+      toast({
+        title: 'ゲストログインではマイページはご利用いただけません',
+        status: 'error',
+        duration: 6000,
+        position: 'top',
+        isClosable: true,
+      });
+    } else {
+      handleClickMyPage();
+    }
+  };
+
   const handleClickMyPage = () => {
     router.push(`/mypage/${userData?.id}`);
   };
+
+  const profileAvatar =
+    isLoggedIn && avatar_url ? avatar_url : '/noNameUser.jpg';
 
   return (
     <Menu>
       <Flex align="center">
         <MenuButton mr={0} px={0} as={Button} bg="none" colorScheme="none">
-          {isLoggedIn && (
-            <HeaderUserIcon src={avatar_url ? avatar_url : '/noNameUser.jpg'} />
-          )}
-          {!isLoggedIn && <HeaderUserIcon src="/noNameUser.jpg" />}
+          <HeaderUserIcon src={profileAvatar} />
         </MenuButton>
       </Flex>
       <MenuList>
-        {isLoggedIn && !isGuest && (
-          <MenuGroup title="Profile">
-            <MenuItem onClick={handleClickMyPage}>
-              <HiUser size="1.5rem" color="gray" />
-              マイページ
-            </MenuItem>
-          </MenuGroup>
-        )}
-        {!isLoggedIn && (
-          <MenuGroup title="Profile">
-            <MenuItem
-              onClick={() =>
-                toast({
-                  title:
-                    'ログインされていない場合、マイページはご利用いただけません',
-                  status: 'error',
-                  duration: 6000,
-                  position: 'top',
-                  isClosable: true,
-                })
-              }
-            >
-              <HiUser size="1.5rem" color="gray" />
-              マイページ
-            </MenuItem>
-          </MenuGroup>
-        )}
-        {isGuest && (
-          <MenuGroup title="Profile">
-            <MenuItem
-              onClick={() =>
-                toast({
-                  title: 'ゲストログインではマイページはご利用いただけません',
-                  status: 'error',
-                  duration: 6000,
-                  position: 'top',
-                  isClosable: true,
-                })
-              }
-            >
-              <HiUser size="1.5rem" color="gray" />
-              マイページ
-            </MenuItem>
-          </MenuGroup>
-        )}
+        <MenuGroup title="Profile">
+          <MenuItem onClick={handleClickUserMenu}>
+            <HiUser size="1.5rem" color="gray" />
+            マイページ
+          </MenuItem>
+        </MenuGroup>
         <MenuDivider />
         <MenuGroup title="other">
           {isLoggedIn && (
